requirements.txt
asgiref==3.8.1
Django==5.1.5
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
isort==6.0.0
phonenumbers==8.13.54
pycountry==24.6.1
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.1
